requirements.txt
hidapi==0.14.0
psutil==7.0.0
keyboard==0.13.5